Add files via upload
Write a Program that should contains
Class Person with fields (name, age, salary, sex)
a static sort function in the class that should take an array of Persons and name of the field and order of sorting and should return a new sorted array based on above inputs. Should not change initial array.
for example Person.sort(arr, 'name', 'asc') -> sort array of persons based on name in ascending order. 'desc' for descending
You have to write a quick sort for this sorting algorithm.